# Turn `FinGenAbGroupElem` into a non-mutable struct