@thofma 2024-09-21
I don't see the harm. I guess many more element objects could/should be immutable.